+++ src/sql/magic.py
@@ -17,6 +17,7 @@ import sql.run
 from sql.store import store
 from sql.command import SQLCommand
 from sql.magic_plot import SqlPlotMagic
+from sql.magic_cmd import SqlCmdMagic
 
 try:
     from traitlets.config.configurable import Configurable
@@ -382,3 +383,4 @@ def load_ipython_extension(ip):
     ip.register_magics(SqlMagic)
     ip.register_magics(RenderMagic)
     ip.register_magics(SqlPlotMagic)
+    ip.register_magics(SqlCmdMagic)